Minor Camera Change

+++ dx11/FrameworkDX11/Camera.cpp
@@ -34,7 +34,7 @@ void Camera::SetView()
 
 void Camera::SetProjection()
 {
-	XMStoreFloat4x4(&_projection, XMMatrixPerspectiveFovLH(XM_PIDIV2, _windowWidth / (FLOAT)_windowHeight, _nearDepth, _farDepth));
+	XMStoreFloat4x4(&_projection, XMMatrixPerspectiveFovLH(XM_PIDIV2, (FLOAT)_windowWidth / (FLOAT)_windowHeight, _nearDepth, _farDepth));
 }
 
 void Camera::Reshape(FLOAT windowWidth, FLOAT windowHeight, FLOAT nearDepth, FLOAT farDepth)
